Add Apache2 license
It has been requested that I add a license so others can have at it.
Sorry this took so long. This project was something I left behind in
favor of using the SCP plugin to copy console logs off host (note that
feature is in the SCP plugin but you ahve to grab a build of latest
master as it hasn't been properl release).